+++ routes/api.php
@@ -1,8 +1,8 @@
 <?php
 
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use TromsFylkestrafikk\Camera\Http\Controllers\CameraController;
+use TromsFylkestrafikk\Camera\Http\Controllers\PictureController;
 
 Route::resource('cameras', CameraController::class)->only(['show']);
-Route::get('cameras/{camera}/file/{file}', [CameraController::class, 'showFile'])->name('camera.file');
+Route::resource('cameras.pictures', PictureController::class)->shallow()->only(['show']);