+++ routes/web.php
@@ -1,0 +1,6 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+use TromsFylkestrafikk\Camera\Http\Controllers\PictureController;
+
+Route::get('pictures/{picture}/download', [PictureController::class, 'download'])->name('camera.picture.download');