+++ src/Jobs/DetectStale.php
@@ -10,7 +10,7 @@ use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Log;
 use TromsFylkestrafikk\Camera\Models\Camera;
-use TromsFylkestrafikk\Camera\Services\CurrentHandler;
+use TromsFylkestrafikk\Camera\Services\CameraService;
 
 /**
  * Detect active cameras with stale imagery.
@@ -36,10 +36,8 @@ class DetectStale implements ShouldQueue
      */
     public function handle()
     {
-        foreach (Camera::isActive()->stale()->get() as $camera) {
-            Log::debug('Camera DetectStale: Found stale camera: ' . $camera->name);
-            $updater = new CurrentHandler($camera);
-            $updater->refresh();
+        foreach (Camera::isActive()->get() as $camera) {
+            CameraService::with($camera)->deactivateIfStalled();
         }
     }
 }