+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Changed
 - Modification of images is now done using pipelines, not .inc files.
   The event `ProcessImage` that this was intended for is also removed.
+- Pictures are now models. This simplified the logic around keeping
+  the latest image up to date a lot. This new model is now the source
+  of broadcasted events, though the channel name is the same.
+- Routes were split in api and web groups, and camera config updated.
 
 ## [0.1.0] – 2021-11-17